Updated file structure to reflect our changes since this branch was made
