+++ core/src/main/se/tevej/game/input/enums/TButton.java
@@ -1,4 +1,4 @@
-package se.tevej.game.input.enums;
+package main.se.tevej.game.input.enums;
 
 public enum TButton {
         KEY_1,
