+++ core/src/main/se/tevej/game/input/listenerInterfaces/OnDraggedListener.java
@@ -1,0 +1,8 @@
+package main.se.tevej.game.input.listenerInterfaces;
+
+import main.se.tevej.game.input.TMouse;
+import main.se.tevej.game.input.enums.TButton;
+
+public interface OnDraggedListener {
+    void onDragged(TMouse mouse, TButton button, float x, float y );
+}
