+++ core/src/main/se/tevej/game/libgdx/view/rendering/input/InputLibgdxAdapter.java
@@ -1,9 +1,9 @@
-package se.tevej.game.libgdx.input;
+package main.se.tevej.game.libgdx.view.rendering.input;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.InputMultiplexer;
 import com.badlogic.gdx.InputProcessor;
-import se.tevej.game.input.enums.TButton;
+import main.se.tevej.game.input.enums.TButton;
 
 import java.util.HashMap;
 import java.util.Map;
